all controller with authentication

+++ TaskMaster/TaskMaster.WebApi/Controllers/FavoritesController.cs
@@ -4,6 +4,7 @@ using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Results;
 using TaskMaster.BLL.MobileService;
+using TaskMaster.BLL.MobileServices;
 using TaskMaster.BLL.WebApiModels;
 
 namespace TaskMaster.WebApi.Controllers
@@ -11,34 +12,47 @@ namespace TaskMaster.WebApi.Controllers
     public class FavoritesController : ApiController
     {
         private readonly FavoritesWebApiService _favoritesWebApiService = new FavoritesWebApiService();
-
+        private readonly VeryficationService _veryficationService = new VeryficationService();
 
         // GET: api/Favorites/email
-        public JsonResult <List<FavoritesMobileDto>>  Get(string email)
+        public JsonResult <List<FavoritesMobileDto>>  Get(string email, string token)
         {
-            var result = _favoritesWebApiService.GetAllFavorites(email);
-            return Json(result);
+            if (_veryficationService.Authorization(email, token))
+            {
+                var result = _favoritesWebApiService.GetAllFavorites(email);
+                return Json(result);
+            }
+            return null;
         }
 
 
         // PUT: api/Favorites
         public HttpResponseMessage Put([FromBody]FavoritesMobileDto favoritesMobileDto)
         {
-            if (_favoritesWebApiService.AddFavorites(favoritesMobileDto))
+            if (_veryficationService.Authorization(favoritesMobileDto.UserEmail, favoritesMobileDto.Token))
             {
-                return new HttpResponseMessage(HttpStatusCode.Accepted);
+                if (_favoritesWebApiService.AddFavorites(favoritesMobileDto))
+                {
+                    return new HttpResponseMessage(HttpStatusCode.Accepted);
+                }
+                return new HttpResponseMessage(HttpStatusCode.BadRequest);
             }
-            return new HttpResponseMessage(HttpStatusCode.BadRequest);
+            return new HttpResponseMessage(HttpStatusCode.Unauthorized);
         }
+
 
         // DELETE: api/Favorites
         public HttpResponseMessage Delete([FromBody]FavoritesMobileDto favoritesMobileDto)
         {
-            if (_favoritesWebApiService.DeleteFromFavorites(favoritesMobileDto))
+            if (_veryficationService.Authorization(favoritesMobileDto.UserEmail, favoritesMobileDto.Token))
             {
-                return new HttpResponseMessage(HttpStatusCode.Accepted);
+                if (_favoritesWebApiService.DeleteFromFavorites(favoritesMobileDto))
+                {
+                    return new HttpResponseMessage(HttpStatusCode.Accepted);
+                }
+                return new HttpResponseMessage(HttpStatusCode.BadRequest);
             }
-            return new HttpResponseMessage(HttpStatusCode.BadRequest);
+            return new HttpResponseMessage(HttpStatusCode.Unauthorized);
         }
     }
 }
